Finished bagofwords.py
Ended with a paragraph describing score comparisons, still running script to add the result .csv to the appropriate subdirectory. A python notebook on this will likely follow shortly. Output .csv is attached, KNN classifiers and regressors had to be abandoned because of repeated memory issues despite troubleshooting in multiple ways.

Signed-off-by: Alec Kulakowski <alecthekulak@gmail.com>